re-divide solver modules; rename analyze.rs => conflict.rs
	renamed:    src/solver/analyze.rs -> src/solver/conflict.rs
	modified:   src/solver/mod.rs
	modified:   src/solver/search.rs
